+++ frontend/src/general/helperFunctions.js
@@ -1,8 +1,19 @@
-
+/**
+ * Allows the path of the history to be changed
+ * This works for all types of histories
+ * @param path
+ * @param history
+ */
 const changePath = (path, history) => {
   history.push(path);
 };
 
+/**
+ * ISO date to string.
+ * The string matches the day in the openhours endpoint
+ * @param date
+ * @returns {string}
+ */
 const getDayForDate = (date) => {
   switch (date.getDay()) {
     case 0:
@@ -24,8 +35,25 @@ const getDayForDate = (date) => {
   }
 };
 
+/**
+ * Generate all the times when a restaurant is opened
+ * @param start
+ * @param end
+ * @returns {[]}
+ */
+const generateAllTimes = (start, end) => {
+  const times = [];
+  for (let i = start; i < end; i += 1) {
+    times.push({
+      time: i > 9 ? `${i}:00:00` : `0${i}:00:00`,
+    });
+  }
+  return times;
+};
+
 export {
   getDayForDate,
+  generateAllTimes,
 };
 
 export default changePath;
